requirements.txt
torch
torchvision
torchaudio
numpy
pillow
opencv-python
ultralytics
transformers
scikit-learn
openvino
scikit-image
